Rename dedup-internals to dedup_internals.
