+++ rlclientlib/rlclientlib.vcxproj.filters
@@ -121,7 +121,7 @@
     <ClInclude Include="serialization\payload_serializer.h" />
     <ClInclude Include="generic_event.h" />
     <ClInclude Include="dedup.h" />
-    <ClInclude Include="dedup-internals.h" />
+    <ClInclude Include="dedup_internals.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
